Added Unit tests for the ComponentLoaderServiceImpl, Fixed some issues over jar: url resolution. Cleaned up some of the configuration mechanisms in the kernel and the loader
git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@54444 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ agnostic/run.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
-rm -rf apache-tomcat-5.5.26
-tar xzf ../../../tars/apache-tomcat-5.5.26.tar.gz
+if [[ a$1 = 'aclean' ]]
+then
+  echo cleaning
+  rm -rf apache-tomcat-5.5.26
+  tar xzf ../../../tars/apache-tomcat-5.5.26.tar.gz
+fi
 cd apache-tomcat-5.5.26
 unzip -o ../overlay/target/assembly-0.1-SNAPSHOT-kernel-loader.zip
 sh bin/catalina.sh run
